refactor: add `FriendRequest` entity to entities barrel export
