+++ backend/src/core/entities/index.ts
@@ -1,2 +1,3 @@
 export * from './user.entity';
 export * from './session.entity';
+export * from './friend_request.entity';
