+++ backend/src/friend_requests/friend_request.module.ts
@@ -2,8 +2,7 @@ import { Module } from '@nestjs/common';
 import { FriendRequestService } from './friend_request.service';
 import { FriendRequestController } from './friend_request.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FriendRequest } from '../core/entities/friend_request.entity';
-import { User } from '../core/entities/user.entity';
+import { User, FriendRequest } from '../core/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([FriendRequest, User])],
