+++ backend/src/friend_requests/friend_request.service.ts
@@ -1,8 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { FriendRequest } from '../core/entities/friend_request.entity';
 import { DeleteResult, Repository } from 'typeorm';
-import { User } from '../core/entities/user.entity';
+import { User, FriendRequest } from '../core/entities';
 
 @Injectable()
 export class FriendRequestService {
